split up lib.rs
* Move user data related code to `userdata` mod
* Move FFI macro related code to `ffi` mod
* Move version related code to `version` mod
