+++ src/connection.rs
@@ -10,16 +10,18 @@ use rustls::{ClientConnection, ServerConnection};
 use crate::certificate::rustls_certificate;
 use crate::enums::rustls_handshake_kind;
 use crate::error::{map_error, rustls_io_result, rustls_result};
+use crate::ffi::{
+    box_castable, free_box, try_callback, try_mut_from_ptr, try_ref_from_ptr, try_slice,
+    try_slice_mut,
+};
 use crate::io::{
     rustls_read_callback, rustls_write_callback, rustls_write_vectored_callback, CallbackReader,
     CallbackWriter, VectoredCallbackWriter,
 };
 use crate::log::{ensure_log_registered, rustls_log_callback};
+use crate::panic::ffi_panic_boundary;
 use crate::rslice::rustls_str;
-use crate::{
-    box_castable, ffi_panic_boundary, free_box, try_callback, try_mut_from_ptr, try_ref_from_ptr,
-    try_slice, try_slice_mut, userdata_push,
-};
+use crate::userdata::userdata_push;
 use rustls_result::NullParameter;
 
 pub(crate) struct Connection {
