+++ src/panic.rs
@@ -124,7 +124,9 @@ macro_rules! ffi_panic_boundary {
             $($tt)*
         }) {
             Ok(ret) => ret,
-            Err(_) => return $crate::PanicOrDefault::value(),
+            Err(_) => return $crate::panic::PanicOrDefault::value(),
         }
     }
 }
+
+pub(crate) use ffi_panic_boundary;
