+++ src/session.rs
@@ -4,7 +4,7 @@ use libc::{c_int, c_void, size_t};
 
 use crate::error::rustls_result;
 use crate::rslice::rustls_slice_bytes;
-use crate::userdata_get;
+use crate::userdata::userdata_get;
 
 /// Any context information the callback will receive when invoked.
 pub type rustls_session_store_userdata = *mut c_void;
