+++ src/verifier.rs
@@ -14,11 +14,13 @@ use webpki::{ExpirationPolicy, RevocationCheckDepth, UnknownStatusPolicy};
 
 use crate::certificate::rustls_root_cert_store;
 use crate::crypto_provider::{self, rustls_crypto_provider};
-use crate::rustls_result::{self, AlreadyUsed};
-use crate::{
-    box_castable, error, ffi_panic_boundary, free_box, set_boxed_mut_ptr, to_boxed_mut_ptr,
-    try_clone_arc, try_mut_from_ptr, try_mut_from_ptr_ptr, try_slice, try_take,
+use crate::error;
+use crate::ffi::{
+    box_castable, free_box, set_boxed_mut_ptr, to_boxed_mut_ptr, try_clone_arc, try_mut_from_ptr,
+    try_mut_from_ptr_ptr, try_slice, try_take,
 };
+use crate::panic::ffi_panic_boundary;
+use crate::rustls_result::{self, AlreadyUsed};
 
 box_castable! {
     /// A built client certificate verifier that can be provided to a `rustls_server_config_builder`
